Adjust compile options for some packages and fix MPI wrappers

+++ bootstrap.bat
@@ -2,10 +2,13 @@
 setlocal enabledelayedexpansion
 
 rem  Fix error C2226: syntax error: unexpected type 'llvmrem sysrem UnicodeCharRange'
-powershell -command "Set-WinSystemLocale -SystemLocale en-US"
+for /f "tokens=2*" %%a in ('powershell -command "Get-WinSystemLocale" ^| findstr en-US') do set SYSTEM_LOCALE=%%a
+echo Current system locale: %SYSTEM_LOCALE%
+if "%SYSTEM_LOCALE%" neq "en-US" powershell -command "Set-WinSystemLocale -SystemLocale en-US"
 
 rem  https://docs.python.org/3/using/windows.html#removing-the-max-path-limitation
-reg add HKLM\SYSTEM\CurrentControlSet\Control\FileSystem /v LongPathsEnabled /t reg_DWORD /d 1
+reg query HKLM\SYSTEM\CurrentControlSet\Control\FileSystem /v LongPathsEnabled >nul 2>&1
+if "%errorlevel%" neq "0" reg add HKLM\SYSTEM\CurrentControlSet\Control\FileSystem /v LongPathsEnabled /t reg_DWORD /d 1
 
 if "%PROCESSOR_ARCHITECTURE%"=="AMD64" (
     set ARCH=x64
@@ -13,9 +16,9 @@ if "%PROCESSOR_ARCHITECTURE%"=="AMD64" (
     set ARCH=x86
 )
 
-set GIT_ROOT="C:\\Program Files\\Git"
+set GIT_ROOT="C:\Program Files\Git"
 set OLDPATH=%PATH%
-set PATH=%PATH%;%~dp0%ARCH%\bin;%GIT_ROOT%\bin
+set PATH=%PATH%;C:\Python312\Scripts;C:\Python312;%HOME%\.cargo\bin;%~dp0%ARCH%\bin;%GIT_ROOT%\bin
 
 cd /d %~dp0
 mkdir %ARCH%\bin
@@ -30,15 +33,15 @@ call :check_python || goto :end
 call :check_git || goto :end
 call :check_rust || goto :end
 
-rem Dealing with line endings
+echo Dealing with line endings
 rem See https://stackoverflow.com/questions/1967370/git-replacing-lf-with-crlf
 git config --global core.autocrlf input
 git config --global core.autocrlf false
 
-rem speed up git handling if on bash environment
+echo Speed up git handling if on bash environment
 git config --system core.fileMode false
 
-rem Changed the default behavior of Git for Windows
+echo Changed the default behavior of Git for Windows
 git config --global core.ignorecase false
 
 call :check_cmake || goto :end
@@ -102,8 +105,8 @@ if "!nv_gpu!" == "" (
   if "%CUDA_PATH%" == "" (
     set with_cuda=
     echo:
-    echo "Do you want to install CUDA and CUDNN? [yes/y/no/n]"
-    echo "If you do not want, just press Enter to cancel"
+    echo Do you want to install CUDA and CUDNN? [yes/y/no/n]
+    echo If you do not want, just press Enter to cancel
     echo:
     set /p with_cuda=
     if "!with_cuda!"=="yes" goto :install_cuda
@@ -119,10 +122,10 @@ for /f "tokens=1-4 delims=." %%a in ("!CUDA_FULL_VERSION!") do set cuda_major_mi
 for /f "delims=_" %%a in ("!CUDA_FULL_VERSION!") do set cuda_version=%%a
 rem https://docs.nvidia.com/cuda/cuda-installation-guide-microsoft-windows/index.html
 if not exist "cuda_!CUDA_FULL_VERSION!_windows.exe" (
-  echo Downloading CUDA !CUDA_FULL_VERSION!, please wait ...
-  wget https://developer.download.nvidia.com/compute/cuda/!CUDA_VERSION!/local_installers/cuda_!CUDA_FULL_VERSION!_windows.exe
+  echo Downloading CUDA !CUDA_FULL_VERSION!
+  wget  --no-check-certificate https://developer.download.nvidia.com/compute/cuda/!CUDA_VERSION!/local_installers/cuda_!CUDA_FULL_VERSION!_windows.exe
 )
-echo Installing CUDA !CUDA_FULL_VERSION!, please wait ...
+echo Installing CUDA !CUDA_FULL_VERSION!
 cuda_!CUDA_FULL_VERSION!_windows.exe -s cuda_profiler_api_!cuda_major_minor! cudart_!cuda_major_minor! cuobjdump_!cuda_major_minor! cupti_!cuda_major_minor! cuxxfilt_!cuda_major_minor! demo_suite_!cuda_major_minor! nvcc_!cuda_major_minor! nvdisasm_!cuda_major_minor! nvfatbin_!cuda_major_minor! nvjitlink_!cuda_major_minor! nvml_dev_!cuda_major_minor! nvprof_!cuda_major_minor! nvprune_!cuda_major_minor! nvrtc_!cuda_major_minor! nvrtc_dev_!cuda_major_minor! opencl_!cuda_major_minor! visual_profiler_!cuda_major_minor! sanitizer_!cuda_major_minor! thrust_!cuda_major_minor! cublas_!cuda_major_minor! cublas_dev_!cuda_major_minor! cufft_!cuda_major_minor! cufft_dev_!cuda_major_minor! curand_!cuda_major_minor! curand_dev_!cuda_major_minor! cusolver_!cuda_major_minor! cusolver_dev_!cuda_major_minor! cusparse_!cuda_major_minor! cusparse_dev_!cuda_major_minor! npp_!cuda_major_minor! npp_dev_!cuda_major_minor! nvjpeg_!cuda_major_minor! nvjpeg_dev_!cuda_major_minor! occupancy_calculator_!cuda_major_minor!
 echo Done.
 
@@ -130,8 +133,8 @@ set CUDNN_VERSION=9.5.0.50
 rem https://docs.nvidia.com/deeplearning/cudnn/latest/reference/support-matrix.html#support-matrix
 rem https://docs.nvidia.com/deeplearning/cudnn/latest/installation/windows.html
 if not exist "cudnn-windows-x86_64-!CUDNN_VERSION!_cuda!cuda_major!-archive.zip" (
-  echo Downloading CUDNN !CUDNN_VERSION!, please wait ...
-  wget https://developer.download.nvidia.com/compute/cudnn/redist/cudnn/windows-x86_64/cudnn-windows-x86_64-!CUDNN_VERSION!_cuda!cuda_major!-archive.zip
+  echo Downloading CUDNN !CUDNN_VERSION!
+  wget  --no-check-certificate https://developer.download.nvidia.com/compute/cudnn/redist/cudnn/windows-x86_64/cudnn-windows-x86_64-!CUDNN_VERSION!_cuda!cuda_major!-archive.zip
 )
 powershell Expand-Archive -Path cudnn-windows-x86_64-!CUDNN_VERSION!_cuda!cuda_major!-archive.zip -DestinationPath . > nul || pause
 if not exist "C:\Program Files\NVIDIA GPU Computing Toolkit\CUDA\v!cuda_major_minor!\lib\x64" mkdir "C:\Program Files\NVIDIA GPU Computing Toolkit\CUDA\v!cuda_major_minor!\lib\x64"
@@ -155,6 +158,7 @@ for /f "usebackq tokens=*" %%i in (`wmic cpu get name ^| findstr "Intel"`) do (
 if "!intel_cpu!" == "" (
   echo You don't have Intel CPU on your PC, but use Intel OneAPI should be OK.
 )
+set oneapi_version=2024.2.1
 set component_lists=intel.oneapi.win.cpp-dpcpp-common
 set with_basekit=no
 set with_hpckit=no
@@ -163,8 +167,8 @@ rem https://oneapi-src.github.io/oneapi-ci/
 if not exist "!ONEAPI_ROOT!\mkl" (
   set with_mkl=
   echo:
-  echo "Do you want to install Intel oneAPI Math Kernel Library ^(oneMKL^)? [yes/y/no/n]"
-  echo "If you do not want, just press Enter to cancel"
+  echo Do you want to install Intel oneAPI Math Kernel Library ^(oneMKL^)? [yes/y/no/n]
+  echo If you do not want, just press Enter to cancel
   echo:
   set /p with_mkl=
   if "!with_mkl!"=="yes" (
@@ -177,26 +181,39 @@ if not exist "!ONEAPI_ROOT!\mkl" (
   )
 )
 if not exist "!ONEAPI_ROOT!\compiler" (
-  if not exist "w_BaseKit_p_2024.2.1.101_offline.exe" (
-    echo Downloading Intel OneAPI BaseKit
-    wget https://registrationcenter-download.intel.com/akdlm/IRC_NAS/d91caaa0-7306-46ea-a519-79a0423e1903/w_BaseKit_p_2024.2.1.101_offline.exe || goto :end
+  if not exist "w_BaseKit_p_!oneapi_version!.101_offline.exe" (
+    echo Downloading Intel OneAPI BaseKit for Windows
+    wget --no-check-certificate https://registrationcenter-download.intel.com/akdlm/IRC_NAS/d91caaa0-7306-46ea-a519-79a0423e1903/w_BaseKit_p_!oneapi_version!.101_offline.exe || goto :end
+  )
+  rem https://www.intel.com/content/www/us/en/developer/articles/tool/compilers-redistributable-libraries-by-version.html
+  if not exist "w_dpcpp_cpp_runtime_p_!oneapi_version!.1084.exe" (
+    echo Downloading Intel oneAPI DPC++/C++ Compiler Runtime for Windows
+    wget --no-check-certificate https://registrationcenter-download.intel.com/akdlm/IRC_NAS/15a35578-2f9a-4f39-804b-3906e0a5f8fc/w_dpcpp_cpp_runtime_p_!oneapi_version!.1084.exe || goto :end
+  )
+  if not exist "w_ifort_runtime_p_!oneapi_version!.1084.exe" (
+    echo Downloading Intel Fortran Compiler Runtime for Windows* ^(IFX/IFORT^)
+    wget --no-check-certificate https://registrationcenter-download.intel.com/akdlm/IRC_NAS/ea23d696-a77f-4a4a-8996-20d02cdbc48f/w_ifort_runtime_p_!oneapi_version!.1084.exe || goto :end
   )
   set with_basekit=yes
 )
 if "!with_basekit!"=="yes" (
   echo Installing Intel OneAPI BaseKit
-  w_BaseKit_p_2024.2.1.101_offline.exe -a --silent --eula accept --components !component_lists!
+  w_BaseKit_p_!oneapi_version!.101_offline.exe -a --silent --eula accept --components !component_lists!
+  echo Installing Intel oneAPI DPC++/C++ Compiler Runtime for Windows
+  start /wait w_dpcpp_cpp_runtime_p_!oneapi_version!.1084.exe || goto :end
+  echo Installing Intel Fortran Compiler Runtime for Windows* ^(IFX/IFORT^)
+  start /wait w_ifort_runtime_p_!oneapi_version!.1084.exe || goto :end
 )
 if not exist "!ONEAPI_ROOT!\mpi" (
-  if not exist "w_HPCKit_p_2024.2.1.80_offline.exe" (
+  if not exist "w_HPCKit_p_!oneapi_version!.80_offline.exe" (
     echo Downloading Intel OneAPI HPCKit
-    wget https://registrationcenter-download.intel.com/akdlm/IRC_NAS/745e923a-3f85-4e1e-b6dd-637c0e9ccba6/w_HPCKit_p_2024.2.1.80_offline.exe || goto :end
+    wget --no-check-certificate https://registrationcenter-download.intel.com/akdlm/IRC_NAS/745e923a-3f85-4e1e-b6dd-637c0e9ccba6/w_HPCKit_p_!oneapi_version!.80_offline.exe || goto :end
   )
   set with_hpckit=yes
 )
 if "!with_hpckit!"=="yes" (
   echo Installing Intel OneAPI HPCKit
-  w_HPCKit_p_2024.2.1.80_offline.exe -a --silent --eula accept --components intel.oneapi.win.ifort-compiler:intel.oneapi.win.mpi.devel
+  w_HPCKit_p_!oneapi_version!.80_offline.exe -a --silent --eula accept --components intel.oneapi.win.ifort-compiler:intel.oneapi.win.mpi.devel
 )
 exit /b 0
 
@@ -207,7 +224,7 @@ rem ============================================================================
 echo Checking wget whether has been installed
 where wget >nul 2>&1
 if "%errorlevel%" neq "0" (
-  echo Installing wget ...
+  echo Installing wget
   curl -L https://eternallybored.org/misc/wget/1.21.4/%ARCH:x=%/wget.exe -o %~dp0%ARCH%\bin\wget.exe || goto :end
 )
 echo Done.
@@ -238,7 +255,7 @@ echo Checking Python whether has been installed
 set PYTHON_VERSION=3.12.8
 where python >nul 2>&1
 if "%errorlevel%" neq "0" (
-  echo Installing python 3 ...
+  echo Installing python 3
   if "%ARCH%" == "x64" (
     if not exist "python-!PYTHON_VERSION!-amd64.exe" (
       wget --no-check-certificate https://www.python.org/ftp/python/!PYTHON_VERSION!/python-!PYTHON_VERSION!-amd64.exe || goto :end
@@ -251,11 +268,10 @@ if "%errorlevel%" neq "0" (
     start /wait python-!PYTHON_VERSION!.exe InstallAllUsers=1 TargetDir=C:\Python312 PrependPath=1 Include_test=0 || goto :end
   )
 )
-echo Installing 3rd party libraries for Python ...
-set PATH=C:\Python312\Scripts;C:\Python312;%PATH%
+echo Installing 3rd party libraries for Python
 python -m pip install --upgrade pip setuptools
 python -m pip install --upgrade meson pygments pyyaml requests rich yamllint psutil || goto :end
-echo Done
+echo Done.
 exit /b 0
 
 rem ==============================================================================
@@ -266,7 +282,7 @@ echo Checking CMake whether has been installed
 set CMAKE_VERSION=3.31.2
 where cmake >nul 2>&1
 if "%errorlevel%" neq "0" (
-  echo Installing CMake ...
+  echo Installing CMake
   if "%ARCH%" == "x64" (
     if not exist "cmake-!CMAKE_VERSION!-windows-x86_64.msi" (
       wget --no-check-certificate https://github.com/Kitware/CMake/releases/download/v!CMAKE_VERSION!/cmake-!CMAKE_VERSION!-windows-x86_64.msi || goto :end
@@ -279,7 +295,7 @@ if "%errorlevel%" neq "0" (
     start /wait msiexec -i cmake-!CMAKE_VERSION!-windows-i368.msi || goto :end
   )
 )
-echo Done
+echo Done.
 exit /b 0
 
 rem ==============================================================================
@@ -290,7 +306,7 @@ echo Checking Git whether has been installed
 set GIT_VERSION=2.47.1
 where git >nul 2>&1
 if "%errorlevel%" neq "0" (
-  echo Intalling Git
+  echo Intalling Git !GIT_VERSION!
   if "%ARCH%" == "x64" (
     if not exist "Git-!GIT_VERSION!-64-bit.exe" (
       wget --no-check-certificate https://github.com/git-for-windows/git/releases/download/v!GIT_VERSION!.windows.1/Git-!GIT_VERSION!-64-bit.exe || goto :end
@@ -303,14 +319,13 @@ if "%errorlevel%" neq "0" (
     start /wait Git-!GIT_VERSION!-32-bit.exe || goto :end
   )
 )
-echo Done
+popd || goto :end
 exit /b 0
 
 rem ==============================================================================
 rem  Download archive and extract it
 rem ==============================================================================
 :download_extract
-setlocal
 set url=%1
 set dest=%2
 set SH="C:\\Program Files\\Git\\bin\\bash.exe"
@@ -329,13 +344,12 @@ rem ============================================================================
 rem  Add more tools in Git for Windows
 rem ==============================================================================
 :git_bash_add_tools
-setlocal
 echo Adding more tools into Git Bash on Windows
 if exist "!GIT_ROOT!" (
   rem zstd
   if not exist "!GIT_ROOT!\mingw64\bin\zstd.exe" (
     if not exist "zstd-v1.5.6-win!ARCH:x=!.zip" (
-      wget --no-check-certificate https://github.com/facebook/zstd/releases/download/v1.5.6/zstd-v1.5.6-win!ARCH:x=!.zip
+      wget --no-check-certificate https://github.com/facebook/zstd/releases/download/v1.5.6/zstd-v1.5.6-win!ARCH:x=!.zip || goto :end
     )
     if not exist "zstd-v1.5.6-win!ARCH:x=!" (
       powershell Expand-Archive -Path zstd-v1.5.6-win!ARCH:x=!.zip -DestinationPath . > nul || goto :end
@@ -412,7 +426,7 @@ rem ============================================================================
 echo Checking yq whether has been installed
 where yq >nul 2>&1
 if "%errorlevel%" neq "0" (
-  echo Installing yq ...
+  echo Installing yq
   if not exist "yq.exe" (
     if "%PROCESSOR_ARCHITECTURE%"=="AMD64" (
       wget --no-check-certificate https://github.com/mikefarah/yq/releases/download/v4.44.3/yq_windows_amd64.exe -O yq.exe || goto :end
@@ -432,7 +446,7 @@ rem ============================================================================
 echo Checking Rust whether has been installed
 where rustc >nul 2>&1
 if "%errorlevel%" neq "0" (
-  echo Installing Rust ...
+  echo Installing Rust
   if not exist "rustup-init.exe" (
     if "%ARCH%" == "x64" (
       wget --no-check-certificate https://static.rust-lang.org/rustup/dist/x86_64-pc-windows-msvc/rustup-init.exe || goto :end
@@ -441,15 +455,14 @@ if "%errorlevel%" neq "0" (
     )
   )
   start /wait rustup-init.exe || goto :end
-  set PATH=%HOME%\.cargo\bin;%PATH%
   rustup default stable-msvc
 )
-echo Done
+echo Done.
 exit /b 0
 
 :end
 cd /d %~dp0
-rmdir /s /q tmp
+if exist "tmp" rmdir /s /q tmp
 set PATH=%OLDPATH%
 set vsinstall=
 set vswhere=
